feat: 도서 이미지 추가 및 포인트 타입 변경

+++ src/main/java/com/nhnacademy/order_payments/dto/order/UserInfoResponse.java
@@ -28,7 +28,7 @@ public record UserInfoResponse(Long userCreatedId,
                                String email,
                                String gradeName,
                                BigDecimal pointRate,
-                               BigDecimal point,
-                               List<InternalAddressResponse> addresses
-) {
+                               Long point,
+                               List<InternalAddressResponse> addresses) {
+    // 주문 서비스가 결제 화면을 그릴 때 원하는 정보만 담기 위해 필요한 DTO
 }
